web/flask/requirements.txt
Flask>=1.1.2
Flask-Bootstrap>=3.3.7.1
waitress>=1.4.3
tensorflow-cpu==2.1.0
numpy>=1.18.3
requests>=2.23.0
Pillow==7.1.0
gunicorn==20.0.4
Flask-SQLAlchemy==2.4.3
Flask-Login==0.5.0
pandas==1.0.5
grpcio==1.31.0
tensorflow-serving-api==2.1.0
